Fixes #20. Handle deletion of one element of a set. (#21)
Consider a set of records using the Weighted routing policy. There are
multiple records with the same name but different SetIdentifier values. If a
setIdentifier is provided in to delRecord, only that particular record should
be deleted; not all with the given name. This patch ensures that if a
setIdentifier is provided to delRecord, only that record in the set will be
deleted.